LayoutTests/http/wpt/web-animations/interfaces/KeyframeEffect/processing-a-keyframes-argument-001-expected.txt


PASS non-animatable property 'animation' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationDelay' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationDirection' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationDuration' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationFillMode' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationIterationCount' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationName' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationPlayState' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animationTimingFunction' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'transition' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'transitionDelay' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'transitionDuration' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'transitionProperty' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'transitionTimingFunction' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'display' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'unsupportedProperty' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'font-size' is not accessed when using a property-indexed keyframe object 
PASS non-animatable property 'animation' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationDelay' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationDirection' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationDuration' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationFillMode' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationIterationCount' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationName' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationPlayState' is not accessed when using a keyframe sequence 
PASS non-animatable property 'animationTimingFunction' is not accessed when using a keyframe sequence 
PASS non-animatable property 'transition' is not accessed when using a keyframe sequence 
PASS non-animatable property 'transitionDelay' is not accessed when using a keyframe sequence 
PASS non-animatable property 'transitionDuration' is not accessed when using a keyframe sequence 
PASS non-animatable property 'transitionProperty' is not accessed when using a keyframe sequence 
PASS non-animatable property 'transitionTimingFunction' is not accessed when using a keyframe sequence 
PASS non-animatable property 'display' is not accessed when using a keyframe sequence 
PASS non-animatable property 'unsupportedProperty' is not accessed when using a keyframe sequence 
PASS non-animatable property 'font-size' is not accessed when using a keyframe sequence 
PASS Equivalent property-indexed and sequenced keyframes: two properties with one value 
PASS Equivalent property-indexed and sequenced keyframes: two properties with three values 
PASS Equivalent property-indexed and sequenced keyframes: two properties with different numbers of values 
PASS Equivalent property-indexed and sequenced keyframes: same easing applied to all keyframes 
PASS Equivalent property-indexed and sequenced keyframes: same composite applied to all keyframes 
FAIL Keyframes are read from a custom iterator assert_equals: number of frames expected 3 but got 0
FAIL 'easing' and 'offset' are ignored on iterable objects assert_equals: number of frames expected 3 but got 0
FAIL Keyframes are read from a custom iterator with multiple properties specified assert_equals: number of frames expected 3 but got 0
FAIL Keyframes are read from a custom iterator with where an offset is specified assert_equals: number of frames expected 3 but got 0
FAIL Reading from a custom iterator that returns a non-object keyframe should throw assert_throws: function "() => {
    new KeyframeEffect(null, createIterable([
      { done: false, value: { left: '100px' } },
      { done: false, value: 1234 },
      { done: false, value: { left: '200px' } },
      { done: true },
    ]));
  }" did not throw
FAIL A list of values returned from a custom iterator should be ignored assert_equals: number of frames expected 1 but got 0
FAIL Only enumerable properties on keyframes are read assert_equals: number of frames expected 2 but got 0
FAIL Only properties defined directly on keyframes are read assert_equals: number of frames expected 2 but got 0
FAIL Only enumerable properties on property-indexed keyframes are read assert_equals: number of frames expected 2 but got 0
FAIL Only properties defined directly on property-indexed keyframes are read assert_equals: number of frames expected 2 but got 0
FAIL Properties are read in ascending order by Unicode codepoint assert_array_equals: property access order property 0, expected "composite" but got "marginLeft"
